+++ README.md
@@ -6,7 +6,7 @@
     - 예약자명: null, ""
     - 날짜: null, "", yyyy-MM-dd이 아닌 경우
     - 시간: null, "", HH:mm이 아닌 경우
-  - [ ] 특정 시간에 대한 예약이 존재하는데, 그 시간을 삭제하려 할 때
+  - [x] 특정 시간에 대한 예약이 존재하는데, 그 시간을 삭제하려 할 때
   - [x] 존재하지 않는 id를 가진 데이터에 접근하려 할 때
   - [x] 지나간 날짜와 시간에 대한 예약 생성 불가능
   - [ ] 중복 예약 불가능
