+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -3,16 +3,20 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.controller.request.ReservationTimeRequest;
+import roomescape.exception.BadRequestException;
 import roomescape.exception.NotFoundException;
 import roomescape.model.ReservationTime;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
 
+    private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationRepository reservationRepository, ReservationTimeRepository reservationTimeRepository) {
+        this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
     }
 
@@ -33,6 +37,10 @@ public class ReservationTimeService {
         Long count = reservationTimeRepository.countReservationTimeById(id);
         if (count == null || count <= 0) {
             throw new NotFoundException("[ERROR] 존재하지 않는 시간입니다.");
+        }
+        Long countOfReservationUsingTime = reservationRepository.countReservationByTimeId(id);
+        if (countOfReservationUsingTime == null || countOfReservationUsingTime > 0) {
+            throw new BadRequestException("[ERROR] 해당 시간을 사용하고 있는 예약이 있습니다.");
         }
         reservationTimeRepository.deleteReservationTime(id);
     }
