+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -7,13 +7,17 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.controller.request.ReservationTimeRequest;
+import roomescape.exception.BadRequestException;
 import roomescape.exception.NotFoundException;
 import roomescape.model.ReservationTime;
 
 import static org.assertj.core.api.Assertions.*;
 
 class ReservationTimeServiceTest {
-    ReservationTimeService reservationTimeService = new ReservationTimeService(new FakeReservationTimeRepository());
+    ReservationTimeService reservationTimeService = new ReservationTimeService(
+            new FakeReservationRepository(),
+            new FakeReservationTimeRepository()
+    );
 
     @DisplayName("모든 예약 시간을 반환한다")
     @Test
@@ -59,5 +63,13 @@ class ReservationTimeServiceTest {
     void should_not_throw_exception_when_exist_id() {
         assertThatCode(() -> reservationTimeService.deleteReservationTime(1))
                 .doesNotThrowAnyException();
+    }
+
+    @DisplayName("특정 시간에 대핸 예약이 존재하는데, 그 시간을 삭제하려 할 때 예외가 발생한다.")
+    @Test
+    void should_throw_exception_when_exist_reservation_using_time() {
+        assertThatThrownBy(() -> reservationTimeService.deleteReservationTime(1))
+                .isInstanceOf(BadRequestException.class)
+                .hasMessage("[ERROR] 해당 시간을 사용하고 있는 예약이 있습니다.");
     }
 }
